+++ lib/queries/collection_object/filter.rb
@@ -295,13 +295,15 @@ module Queries
       # See Queries::ColletingEvent::Filter for other use
       def determiner_facet
         return nil if determiner_id.empty?
+        tt = table
+
         o = ::TaxonDetermination.arel_table
         r = ::Role.arel_table
 
-        a = o.alias("a_")
+        a = o.alias("a_det__")
         b = o.project(a[Arel.star]).from(a)
 
-        c = r.alias('r1')
+        c = r.alias('det_r1')
 
         b = b.join(c, Arel::Nodes::OuterJoin)
           .on(
@@ -316,9 +318,10 @@ module Queries
         b = b.where(e.and(f))
         b = b.group(a['id'])
         b = b.having(a['id'].count.eq(determiner_id.length)) unless determiner_id_or
+
         b = b.as('det_z1_')
 
-        ::CollectionObject.joins(:taxon_determinations).joins(Arel::Nodes::InnerJoin.new(b, Arel::Nodes::On.new(b['id'].eq(o['id']))))
+        ::CollectionObject.joins(Arel::Nodes::InnerJoin.new(b, Arel::Nodes::On.new(b['biological_collection_object_id'].eq(tt['id']))))
       end
 
       def georeferences_facet